+++ src/db/components_db_data.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace vle_solvers
+{
+    extern const char* thermo_db_serialized;
+}
