+++ src/vle_solvers.h
@@ -1,6 +1,13 @@
 #pragma once
 
-#include "db/common_db.h"
-#include "db/components_db.h"
-#include "db/tdb_serialize_2024_05_24.h"
-#include "db/tdb_serialize_common.h"
+#ifndef VLE_SOLVER_H
+#define VLE_SOLVER_H
+
+#include "src/common_db.h"
+#include "src/components_db.h"
+#include "src/components_db_data.h"
+#include "src/components_db_init.h"
+#include "src/tdb_serialize_2024_05_24.h"
+#include "src/tdb_serialize_common.h"
+
+#endif // VLE_SOLVER_H
